Move LabelerDecarlation to Labeler namespace, and rename to Service. Split up Labeler methods into individual files.
