+++ src/idunno.Bluesky/Labeler/Defs/LabelerPolicies.cs
@@ -3,7 +3,9 @@
 
 using idunno.AtProto.Labels;
 
+#pragma warning disable IDE0130 // Namespace does not match folder structure
 namespace idunno.Bluesky.Labeler;
+#pragma warning restore IDE0130 // Namespace does not match folder structure
 
 /// <summary>
 /// The policies for a labeler
