+++ src/idunno.Bluesky/Labeler/Defs/LabelerView.cs
@@ -7,7 +7,9 @@ using idunno.AtProto;
 using idunno.AtProto.Labels;
 using idunno.Bluesky.Actor;
 
+#pragma warning disable IDE0130 // Namespace does not match folder structure
 namespace idunno.Bluesky.Labeler;
+#pragma warning restore IDE0130 // Namespace does not match folder structure
 
 /// <summary>
 /// Encapsulates a detailed view over a labeller.
