+++ src/idunno.Bluesky/Labeler/Defs/LabelerViewDetailed.cs
@@ -3,7 +3,9 @@
 
 using idunno.AtProto;
 
+#pragma warning disable IDE0130 // Namespace does not match folder structure
 namespace idunno.Bluesky.Labeler;
+#pragma warning restore IDE0130 // Namespace does not match folder structure
 
 /// <summary>
 /// Encapsulates a detailed view over a labeler.
